+++ pkg/provider/util_test.go
@@ -92,6 +92,70 @@ func TestHasComputedValue(t *testing.T) {
 			}},
 			hasComputedValue: true,
 		},
+		{
+			name: "Object with nested slice of map[string]interface{} has a computed value",
+			obj: &unstructured.Unstructured{Object: map[string]interface{}{
+				"field1": 1,
+				"field2": []map[string]interface{}{
+					{"field3": resource.Computed{}},
+				},
+			}},
+			hasComputedValue: true,
+		},
+		{
+			name: "Object with nested slice of interface{} has a computed value",
+			obj: &unstructured.Unstructured{Object: map[string]interface{}{
+				"field1": 1,
+				"field2": []interface{}{
+					resource.Computed{},
+				},
+			}},
+			hasComputedValue: true,
+		},
+		{
+			name: "Object with nested slice of map[string]interface{} with nested slice of interface{} has a computed value",
+			obj: &unstructured.Unstructured{Object: map[string]interface{}{
+				"field1": 1,
+				"field2": []map[string]interface{}{
+					{"field3": []interface{}{
+						resource.Computed{},
+					}},
+				},
+			}},
+			hasComputedValue: true,
+		},
+		{
+			name: "Complex nested object with computed value",
+			obj: &unstructured.Unstructured{Object: map[string]interface{}{
+				"field1": 1,
+				"field2": []map[string]interface{}{
+					{"field3": []interface{}{
+						[]map[string]interface{}{
+							{"field4": []interface{}{
+								resource.Computed{},
+							}},
+						},
+					}},
+				},
+			}},
+			hasComputedValue: true,
+		},
+		{
+			name: "Complex nested object with no computed value",
+			obj: &unstructured.Unstructured{Object: map[string]interface{}{
+				"field1": 1,
+				"field2": []map[string]interface{}{
+					{"field3": []interface{}{
+						[]map[string]interface{}{
+							{"field4": []interface{}{
+								"field5",
+							}},
+						},
+					}},
+				},
+			}},
+			hasComputedValue: false,
+		},
 	}
 
 	for _, test := range tests {